Fix bug where editing/deleting a new coin doesn't save changes
The issue is due to the fact new Coins have coinId=0, so when we try to
assign the newly created coin to CoinDetailFragment.coin, it doesn't
have a valid id. This change exposes the id, returned from the insert
into the local db, and saves it in a field in the ViewModel.

This may have introduced a race condition. I'll have to watch it (and
write tests to cover this scenario). I need to investigate if I can just
return it from the ViewModel.