"Fix" race condition saving coin to database
Saving and re-editing a new coin before going back to the listview
wouldn't get saved (but would under the debugger). This implied a race
condition setting the viewmodel's lastCoinId and the listview reading
it. This change synchronizes the coroutine calls, so it now works, but
certainly isn't the best solution. There's got to be a better way!